// src/app/(entities)/character/character.types.ts
export type ICharacterResponse = {
    count: number;
    next: string | null;
    previous: string | null;
    results: ICharacter[];
  };

export type ICharacter = {
    name: string;
    height: string;
    mass: string;
    hair_color: string;
    skin_color: string;
    eye_color: string;
    birth_year: string;
    gender: string;
    homeworld: string;
    films: string[];
    species: string[];
    vehicles: string[];
    starships: string[];
    created: string;
    edited: string;
    url: string;
  };